Another fix for an old migration
this one is not minor though XD When recreating the table, the order
of the columns was different, which means, doing a 'select *' meant
that the `created_at` column values where put into the `visits`
column. It is a wounder nothing important got blown up XD

I had to fix the database manually.